api/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.88.0"
pydantic = "^1.10.2"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
SQLAlchemy = "^1.4.31"
asyncpg = "^0.27.0"
alembic = "^1.7.6"
python-dotenv = "^0.19.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
cryptography = "^38.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pytest = "^7.0.1"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110
extend-exclude = "src/db/migrations"
target-version = ["py39"]

[tool.ruff]
line-length = 110
fix = true
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "N",   # pep8-naming
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "C90", # mccabe
    "ERA", # eradicate
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "N805",  # first argument of a method should be named self
    "S106",  # possible hardcoded password
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
